+++ account/User.py
@@ -1,8 +1,13 @@
 class User:
+
+    count = 0
+
     def __init__(self, email, password, admin=False):
         self.__email = email
         self.__password = password
         self.__admin = admin
+        self.__class__.count += 1
+        self.__accountId = self.__class__.count
 
     def set_email(self, email):
         self.__email = email
@@ -18,3 +23,7 @@ class User:
 
     def get_admin(self):
         return self.__admin
+
+    def get_accountId(self):
+        return self.__accountId
+
