+++ templates/includes/_navbar.html
@@ -43,7 +43,7 @@
               <!-- Text based logo -->
               <a href="/index">
                 <span class="fa fa-shopping-cart"></span>
-                <p><strong>Fugl</strong> Mart<span>The finest groceries with us.</span></p>
+                <p><strong>Fugl</strong> Store<span>The finest groceries with us.</span></p>
               </a>
               <!-- img based logo -->
               <!-- <a href="index"><img src="img/logo.jpg" alt="logo img"></a> -->
